Update module for DBIC provider
In order to add DBIC support, changes to the overall plugin were needed.
This was so that the DBIC schema could be retrieved using $dsl->schema.
The schema has to be captured on module load, otherwise it's not
available in later calls.

Without version information, Dancer throws errors when trying to print
the version information on module load.

The DBIC provider is located at:
https://github.com/ctrlo/Dancer2-Plugin-Auth-Extensible-Provider-DBIC