Fix failed to send return when user not found during PW reset
In the event of a user not being found (unable to set user details) the
user_password function was returning undef (failed to send email) rather
than 0, for no emails sent. This fixes that.

It uses the "sometimes to be avoided" $@ return value with eval, but I
believe it is fine in this circumstance.