use Dancer2::Plugin::Database in database provider
Providers are independent classes which are used by the Auth::Extensible
plugin but are not in themselves part of the plugin so stop pretending
that they are and instead just make use of the `database` keyword.

This simple change makes the database provider compatible with current
Dancer release (v0.166001) as well as current master which includes
plugin2 code.