change_session_id on successful login if available
This is a method that has not yet been merged into Dancer2 core but I
expect it to be there soon.

See https://github.com/PerlDancer/Dancer2/pull/1195

It is considered good security practice to change session id on any
change of privilege level (for example on login):

https://www.owasp.org/index.php/Session_Management_Cheat_Sheet#Renew_the_Session_ID_After_Any_Privilege_Level_Change

This is also a requirement of a number of established security
standards.