login_template config attr should not have a true default
Previously the default was 'login' which meant that if the user had a
template called 'login' then that would be used which might not be what
a user expects. Instead make the default an empty string so the user
must set this config value in order to use a custom template.

Also some code logic improvements + fixes for this change.