Load HTTP::BrowserDetect at compile time
Before this commit, HTTP::BrowserDetect was loaded at runtime. The
problem with this was that if it wasn't installed for some reason, then
tests would fail with a mysterious 500 error message, and applications
would presumably suffer a similar fate.

I'm not sure of the original reason for loading at runtime, so if this
change causes you any problems, then please raise a bug report.